디자인 및 전반적인 사용성 수정
