+++ src/components/Card.js
@@ -4,52 +4,7 @@ import MDEditor from '@uiw/react-md-editor';
 
 import Styles from './Card.module.css';
 import { useNavigate } from 'react-router-dom';
-
-function SimpleDateTimeFormat(date, pattern) {
-	var dateString = pattern.replace(/(yyyy|MM|dd|HH|mm|ss|SSS)/g, function(match) {
-		var matchString = "";
-		switch(match) {
-			case "yyyy":
-				matchString = date.getFullYear();
-				break;
-			case "MM":
-				matchString = date.getMonth() + 1;
-				break;
-			case "dd":
-				matchString = date.getDate();
-				break;
-			case "HH":
-				matchString = date.getHours();
-				break;
-			case "mm":
-				matchString = date.getMinutes();
-				break;
-			case "ss":
-				matchString = date.getSeconds();
-				break;
-			case "SSS":
-				matchString = date.getMilliseconds();
-				break;
-			default :
-				matchString = match;
-				break;
-		}
-		if (match == "SSS") {
-			if (matchString < 10) {
-				matchString = "00" + matchString;
-			} else if (matchString < 100) {
-				matchString = "0" + matchString;
-			}
-		} else {
-			if ((typeof(matchString) == "number" && matchString < 10)) {
-				matchString = "0" + matchString;
-			}
-		}
-		return matchString;
-	});
-
-	return dateString;
-}
+import DateFunction from './DateFunction';
 
 const Card = (props) => {
 
@@ -61,10 +16,10 @@ const Card = (props) => {
 
     const navigate = useNavigate();
 
-    let written = SimpleDateTimeFormat(new Date(card.writtenDate), "yyyy.MM.dd HH:mm");
+    let written = DateFunction(new Date(card.writtenDate), "MM.dd HH:mm");
 
     if(card.isYours) {
-        let review = SimpleDateTimeFormat(new Date(card.latestReviewDate), "yyyy.MM.dd HH:mm");
+        let review = DateFunction(new Date(card.latestReviewDate), "MM.dd HH:mm");
         return (
             <div 
                 style={{
