+++ src/components/Card.module.css
@@ -1,6 +1,6 @@
 .cardDiv {
-    border: 1.5px black solid;
-    border-radius: 10px;
+    border: 0.2px #00000033 solid;
+    border-radius: 20px;
     display: flex;
     flex-direction: column;
     cursor: pointer;
@@ -25,8 +25,18 @@
     padding: 20px;
     height: 200px;
     overflow-y: scroll;
-    border-radius: 10px;
-    background-color: #D9D9D9AA;
+    border-radius: 5px;
+    scrollbar-width: 1px;
+    background-color: #D9D9D944;
+}
+
+.articleDiv::-webkit-scrollbar {
+    width: 5px;
+}
+
+.articleDiv::-webkit-scrollbar-thumb {
+    background: #666;
+    border-radius: 2.5px;
 }
 
 .hashtagsDiv {
@@ -48,11 +58,13 @@
 .statDiv {
     flex-direction: row;
     display: flex;
-    padding: 0px 10px 0px 10px;
+    padding: 0px 20px 0px 20px;
     justify-content: space-between;
+    align-items: center;
     font-size: 12px;
 }
 
 .statIcon {
     text-align: center;
+    align-self: center;
 }
