+++ src/screen/main/CardEditing.module.css
@@ -13,6 +13,7 @@
 
 .writingDiv {
     width: 700px;
+    height: 80%;
     background-color: white;
     padding: 20px;
     border-radius: 10px;
@@ -45,15 +46,12 @@
 
 .editorDiv {
     margin: 10px 0px 0px 0px;
+    height: 100%;
 }
 
 .katexGuide {
     font-size: 12px;
     font-style: italic;
-}
-
-.editor {
-    
 }
 
 .bottomDiv {
