+++ src/screen/main/CardReading.js
@@ -10,52 +10,7 @@ import { useEffect } from 'react';
 import MDEditor from '@uiw/react-md-editor';
 
 import Backend from '../../axios/Backend';
-
-function SimpleDateTimeFormat(date, pattern) {
-	var dateString = pattern.replace(/(yyyy|MM|dd|HH|mm|ss|SSS)/g, function(match) {
-		var matchString = "";
-		switch(match) {
-			case "yyyy":
-				matchString = date.getFullYear();
-				break;
-			case "MM":
-				matchString = date.getMonth() + 1;
-				break;
-			case "dd":
-				matchString = date.getDate();
-				break;
-			case "HH":
-				matchString = date.getHours();
-				break;
-			case "mm":
-				matchString = date.getMinutes();
-				break;
-			case "ss":
-				matchString = date.getSeconds();
-				break;
-			case "SSS":
-				matchString = date.getMilliseconds();
-				break;
-			default :
-				matchString = match;
-				break;
-		}
-		if (match == "SSS") {
-			if (matchString < 10) {
-				matchString = "00" + matchString;
-			} else if (matchString < 100) {
-				matchString = "0" + matchString;
-			}
-		} else {
-			if ((typeof(matchString) == "number" && matchString < 10)) {
-				matchString = "0" + matchString;
-			}
-		}
-		return matchString;
-	});
-
-	return dateString;
-}
+import DateFunction from '../../components/DateFunction';
 
 const CardReading = () => {
 
@@ -88,7 +43,6 @@ const CardReading = () => {
         }
 
         data = data.data;
-        console.log(data);
         setCard(data);
     }
 
@@ -175,12 +129,12 @@ const CardReading = () => {
                         ?
                         <div className={Styles.infoDiv}>
                             <p className={Styles.infoText}>{card.reviewCount} 회 복습</p>
-                            <p className={Styles.infoText}><Create fontSize='12px' /> {SimpleDateTimeFormat(new Date(card.writtenDate), "yyyy.MM.dd HH:mm")}</p>
-                            <p className={Styles.infoText}><RateReview fontSize='12px' /> {SimpleDateTimeFormat(new Date(card.latestReviewDate), "yyyy.MM.dd HH:mm")}</p>
+                            <p className={Styles.infoText}><Create fontSize='12px' /> {DateFunction(new Date(card.writtenDate), "yyyy.MM.dd HH:mm")}</p>
+                            <p className={Styles.infoText}><RateReview fontSize='12px' /> {DateFunction(new Date(card.latestReviewDate), "yyyy.MM.dd HH:mm")}</p>
                         </div>
                         :
                         <div className={Styles.infoDiv}>
-                            <p><Create fontSize='12px' /> {SimpleDateTimeFormat(new Date(card.writtenDate), "yyyy.MM.dd HH:mm")}</p>
+                            <p><Create fontSize='12px' /> {DateFunction(new Date(card.writtenDate), "yyyy.MM.dd HH:mm")}</p>
                         </div>
                     }
             </div>
