+++ src/screen/main/CardReading.module.css
@@ -23,6 +23,7 @@
     background-color: white;
     padding: 0px 20px 20px 20px;
     border-radius: 10px;
+    height: 80%;
 }
 
 .topDiv {
@@ -56,6 +57,17 @@
 
 .contentDiv {
     margin-bottom: 20px;
+    height: 80%;
+    overflow: scroll;
+}
+
+.contentDiv::-webkit-scrollbar {
+    width: 5px;
+}
+
+.contentDiv::-webkit-scrollbar-thumb {
+    background: #666;
+    border-radius: 2.5px;
 }
 
 .infoDiv {
@@ -95,13 +107,13 @@
     align-items: center;
     flex-wrap: wrap;
     margin: 7px 0px;
+    height: 20px;
 }
 
 .bottomHashtagItem {
     border-radius: 8px;
     margin: 0px 10px 0px 0px;
     font-weight: bold;
-    cursor: pointer;
     font-size: 15px;
     line-height: 20px;
 }
@@ -117,7 +129,6 @@
     padding: 0px 15px;
     font-size: 20px;
     font-weight: bold;
-    cursor: pointer;
     width: 100px;
     flex-shrink: 0;
 }
